+++ apps/web/prisma/schema.prisma
@@ -25,6 +25,7 @@ model Post {
   thumbnailUrl String
   votesCount   Int         @default(0)
   hasAi        Boolean     @default(false)
+  /// if the post is deleted, it means it was deleted by the user or product hunt team
   deleted      Boolean     @default(false)
   topics       TopicPost[]
 }
