+++ apps/web/src/app/api/daily-email/route.ts
@@ -1,7 +1,7 @@
 import env from "@/app/env";
-import { getYesterdaysLaunches } from "@/app/lib/persistence";
-import type { NextRequest } from "next/server";
+import { getYesterdaysLaunches } from "@/app/lib/launches";
 import { LoopsClient } from "loops";
+import type { NextRequest } from "next/server";
 
 const loops = new LoopsClient(env.LOOPS_API_KEY);
 
@@ -19,19 +19,20 @@ export async function GET(request: NextRequest) {
   const topThree = launches.slice(0, 3);
 
   // Map the launches to the required event properties schema
-  const eventProperties = topThree.reduce((acc, launch, index) => {
-    const num = index + 1;
-    acc[`product_${num}.title`] = launch.name || `Test${num}`;
-    acc[`product_${num}.desc`] = launch.description || "something cool";
-    return acc;
-  }, {} as Record<string, string>);
-
+  const eventProperties = topThree.reduce(
+    (acc, launch, index) => {
+      const num = index + 1;
+      acc[`product_${num}.title`] = launch.name || `Test${num}`;
+      acc[`product_${num}.desc`] = launch.description || "something cool";
+      return acc;
+    },
+    {} as Record<string, string>,
+  );
 
   // const event = await loops.sendEvent({
   //   eventName: "daily_launches",
   //   eventProperties,
   // });
 
-  return Response.json({ eventProperties});
-
-} 
+  return Response.json({ eventProperties });
+}
