+++ apps/web/src/app/api/stats/route.ts
@@ -2,7 +2,7 @@ import type { ProductPost } from "@/app/types";
 import { getStartAndEndOfDayInUTC } from "@/app/utils/date";
 import { filterPosts } from "@/app/utils/string";
 import { unstable_cache } from "next/cache";
-import { getTodaysLaunches } from "../../lib/persistence";
+import { getTodaysLaunches } from "../../lib/launches";
 
 export const dynamic = "force-dynamic";
